configuration fixes

+++ ForcePlatformCore/Models/AppsettingsModel.cs
@@ -11,7 +11,7 @@ namespace WindowsFormsApp1.Models
         public bool AutoSelectCom { get; set; }
         public string ComPort { get; set; }
         public int FilterLength { get; set; }
-        public int CalibrateZ { get; set; }
+        public double CalibrateZ { get; set; }
         public double FreeFallAcc { get; set; }
     }
 }
